feat/objects refactor
